+++ src/app/shared/BibleBooks.ts
@@ -1,0 +1,5 @@
+export interface BibleBooks {
+    bookName: string,
+    bookId: number,
+    subcategory: string,
+}